ecommerce_dbt/ecommerce_streamlit/requirements.txt
streamlit
snowflake-connector-python
pandas
plotly
python-dotenv
